refactor: use intersectionWith func from Data.Map
Use the intersectionWith function from Data.Map for zipping instead of
implementing it with potential errors due to the first map having keys
that are not present in the second map.

If the scenario described above was expected to cause an error, this
commit introduces a regression and should be reverted.