refactor: simplify listDirectoryRecursive
Simplify listDirectoryRecursive by further leveraging the fact that
lists are monads and using Control.Monad.

This isn't a pure refactor of the action since the output IS different:
the previous action would descend the directory tree immediately upon
finding a directory; now the recursive descent is left for last.

This change in behavior might be overcome by using Data.List.sort.